\name{biwt.est.2vecs}
\alias{biwt.est.2vecs}
\title{biwt est 2vecs}
\description{tukey biweight estimation of just two vectors}
\usage{
biwt.est.2vecs(x,
r = 0.2,
med.init = covMcd(x))
}
\arguments{
  \item{x}{matrix with two vectors}
  \item{r}{the breakdown}
  \item{med.init}{init}
}

\value{returns the biweight means and cov matrix.}

\author{David L Gibbs}





