\name{estimateBiWt}
\alias{estimateBiWt}
\title{estimateBiWt}
\description{parallel tukey estimate call}
\usage{
estimateBiWt(isandjs,
r,
med.init,
median)
}
\arguments{
  \item{isandjs}{matrix with 2 columns.}
  \item{r}{the breakdown}
  \item{med.init}{the median init value}
  \item{median}{the flag for median or covMcd}
}

\value{the tukey correlation}

\author{David L Gibbs}





