\name{fastukey-package}
\alias{fastukey-package}
\docType{package}
\title{Performs Tukeys biwt correlation... fast! (see biwt package).}
\description{Extension of the biwt package for speed.}
\details{
\tabular{ll}{Package: \tab fastukey\cr
Imports: \tab biwt, multicore\cr
Type: \tab Package\cr
Title: \tab Performs Tukeys biwt correlation... fast! (see biwt package).\cr
Version: \tab 1.0\cr
Date: \tab 2011-08-26\cr
Author: \tab David L Gibbs\cr
Maintainer: \tab David L Gibbs <gibbsd@ohsu.edu>\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr}
}
\author{David L Gibbs}

\keyword{ package }


