\name{fastukeycor-package}
\alias{fastukeycor-package}
\docType{package}
\title{Performs Tukeys biwt correlation... fast! (see biwt).}
\description{Code and ideas modified from biwt package.}
\details{
\tabular{ll}{Package: \tab fastukeycor\cr
Type: \tab Package\cr
Title: \tab Performs Tukeys biwt correlation... fast! (see biwt).\cr
Version: \tab 1.0\cr
Date: \tab 2011-08-26\cr
Author: \tab David L Gibbs\cr
Maintainer: \tab David L Gibbs <gibbsd@ohsu.edu>\cr
License: \tab GPLv3\cr
LazyLoad: \tab yes\cr}
}
\author{David L Gibbs}

\keyword{ package }


