\name{fullInitFun}
\alias{fullInitFun}
\title{fullInitFun}

\usage{
fullInitFun(x,
full.init,
median)
}
\arguments{
  \item{x}{data matrix}
  \item{full.init}{logic indicator for init using full data matrix}
  \item{median}{logic indicator for median vs covMcd}
}



\author{David L Gibbs}





